Extract initial common code into common directory
