+++ backend/compact-connect/lambdas/common/config.py
@@ -26,7 +26,7 @@ class _Config:
 
     @cached_property
     def data_client(self):
-        from data_model.client import DataClient
+        from common.data_model.client import DataClient
 
         return DataClient(self)
