+++ backend/compact-connect/lambdas/common/data_model/query_paginator.py
@@ -4,11 +4,11 @@ from collections.abc import Callable
 from types import MethodType
 
 from botocore.exceptions import ClientError
-from config import config, logger
-from exceptions import CCInternalException, CCInvalidRequestException
+from common.config import config, logger
+from common.exceptions import CCInternalException, CCInvalidRequestException
 from marshmallow import ValidationError
 
-from data_model.schema.base_record import BaseRecordSchema
+from common.data_model.schema.base_record import BaseRecordSchema
 
 
 # It's conventional to name a decorator in snake_case, even if it is implemented as a class
