+++ backend/compact-connect/lambdas/common/data_model/schema/__init__.py
