+++ backend/compact-connect/lambdas/common/data_model/schema/base_record.py
@@ -4,7 +4,7 @@
 from abc import ABC
 from datetime import UTC, datetime
 
-from exceptions import CCInternalException
+from common.exceptions import CCInternalException
 from marshmallow import EXCLUDE, RAISE, Schema, post_load, pre_dump
 from marshmallow.fields import UUID, Date, List, String
 from marshmallow.validate import Regexp
