+++ backend/compact-connect/lambdas/common/data_model/schema/provider.py
@@ -1,28 +1,28 @@
 # ruff: noqa: N801, N815, ARG002  invalid-name unused-argument
+from urllib.parse import quote
 
-from config import config
-from marshmallow import ValidationError, pre_dump, validates_schema
+from common.config import config
+from marshmallow import ValidationError, post_load, pre_dump, validates_schema
 from marshmallow.fields import UUID, Boolean, Date, Email, String
 from marshmallow.validate import Length, OneOf, Regexp
 
-from data_model.schema.base_record import BaseRecordSchema, ForgivingSchema, ITUTE164PhoneNumber, SocialSecurityNumber
+from common.data_model.schema.base_record import (
+    BaseRecordSchema,
+    ForgivingSchema,
+    ITUTE164PhoneNumber,
+    Set,
+    SocialSecurityNumber,
+)
 
 
-class LicenseCommonSchema(ForgivingSchema):
-    pass
-
-
-class LicensePublicSchema(LicenseCommonSchema):
+class ProviderPublicSchema(ForgivingSchema):
     """Schema for license data that can be shared with the public"""
 
-    birthMonthDay = String(required=False, allow_none=False, validate=Regexp('^[0-1]{1}[0-9]{1}-[0-3]{1}[0-9]{1}'))
-
-
-class LicensePostSchema(LicensePublicSchema):
-    """Schema for license data as posted by a board"""
+    # Provided fields
+    providerId = UUID(required=True, allow_none=False)
 
     compact = String(required=True, allow_none=False, validate=OneOf(config.compacts))
-    jurisdiction = String(required=True, allow_none=False, validate=OneOf(config.jurisdictions))
+    licenseJurisdiction = String(required=True, allow_none=False, validate=OneOf(config.jurisdictions))
     ssn = SocialSecurityNumber(required=True, allow_none=False)
     npi = String(required=False, allow_none=False, validate=Regexp('^[0-9]{10}$'))
     licenseType = String(required=True, allow_none=False)
@@ -31,8 +31,6 @@ class LicensePostSchema(LicensePublicSchema):
     middleName = String(required=False, allow_none=False, validate=Length(1, 100))
     familyName = String(required=True, allow_none=False, validate=Length(1, 100))
     suffix = String(required=False, allow_none=False, validate=Length(1, 100))
-    dateOfIssuance = Date(required=True, allow_none=False)
-    dateOfRenewal = Date(required=True, allow_none=False)
     dateOfExpiration = Date(required=True, allow_none=False)
     dateOfBirth = Date(required=True, allow_none=False)
     homeAddressStreet1 = String(required=True, allow_none=False, validate=Length(2, 100))
@@ -44,6 +42,9 @@ class LicensePostSchema(LicensePublicSchema):
     emailAddress = Email(required=False, allow_none=False, validate=Length(1, 100))
     phoneNumber = ITUTE164PhoneNumber(required=False, allow_none=False)
 
+    # Generated fields
+    birthMonthDay = String(required=False, allow_none=False, validate=Regexp('^[0-1]{1}[0-9]{1}-[0-3]{1}[0-9]{1}'))
+
     @validates_schema
     def validate_license_type(self, data, **kwargs):  # noqa: ARG001 unused-argument
         license_types = config.license_types_for_compact(data['compact'])
@@ -51,17 +52,46 @@ class LicensePostSchema(LicensePublicSchema):
             raise ValidationError({'licenseType': f"'licenseType' must be one of {license_types}"})
 
 
-@BaseRecordSchema.register_schema('license')
-class LicenseRecordSchema(BaseRecordSchema, LicensePostSchema):
+@BaseRecordSchema.register_schema('provider')
+class ProviderRecordSchema(BaseRecordSchema, ProviderPublicSchema):
     """Schema for license records in the license data table"""
 
-    _record_type = 'license'
+    _record_type = 'provider'
 
-    # Provided fields
-    providerId = UUID(required=True, allow_none=False)
+    # Generated fields
+    privilegeJurisdictions = Set(String, required=False, allow_none=False, load_default=set())
+    providerFamGivMid = String(required=False, allow_none=False, validate=Length(2, 400))
+    providerDateOfUpdate = Date(required=True, allow_none=False)
 
     @pre_dump
     def generate_pk_sk(self, in_data, **kwargs):  # noqa: ARG001 unused-argument
         in_data['pk'] = f'{in_data['compact']}#PROVIDER#{in_data['providerId']}'
-        in_data['sk'] = f'{in_data['compact']}#PROVIDER#license/{in_data['jurisdiction']}'
+        in_data['sk'] = f'{in_data['compact']}#PROVIDER'
+        return in_data
+
+    @pre_dump
+    def populate_birth_month_day(self, in_data, **kwargs):  # noqa: ARG001 unused-argument
+        in_data['birthMonthDay'] = in_data['dateOfBirth'].strftime('%m-%d')
+        return in_data
+
+    @pre_dump
+    def populate_prov_date_of_update(self, in_data, **kwargs):  # noqa: ARG001 unused-argument
+        in_data['providerDateOfUpdate'] = in_data['dateOfUpdate']
+        return in_data
+
+    @post_load
+    def drop_prov_date_of_update(self, in_data, **kwargs):  # noqa: ARG001 unused-argument
+        del in_data['providerDateOfUpdate']
+        return in_data
+
+    @pre_dump
+    def populate_fam_giv_mid(self, in_data, **kwargs):  # noqa: ARG001 unused-argument
+        in_data['providerFamGivMid'] = '#'.join(
+            (quote(in_data['familyName']), quote(in_data['givenName']), quote(in_data.get('middleName', ''))),
+        )
+        return in_data
+
+    @post_load
+    def drop_fam_giv_mid(self, in_data, **kwargs):  # noqa: ARG001 unused-argument
+        del in_data['providerFamGivMid']
         return in_data
