+++ backend/compact-connect/lambdas/common/utils.py
@@ -8,8 +8,8 @@ from uuid import UUID
 
 from aws_lambda_powertools.utilities.typing import LambdaContext
 from botocore.exceptions import ClientError
-from config import logger
-from exceptions import CCAccessDeniedException, CCInvalidRequestException, CCNotFoundException, CCUnauthorizedException
+from common.config import logger
+from common.exceptions import CCAccessDeniedException, CCInvalidRequestException, CCNotFoundException, CCUnauthorizedException
 
 
 class ResponseEncoder(JSONEncoder):
