+++ backend/compact-connect/lambdas/provider-data-v1/handlers/providers.py
@@ -1,10 +1,10 @@
 import json
 
 from aws_lambda_powertools.utilities.typing import LambdaContext
-from config import config, logger
-from exceptions import CCInvalidRequestException
+from common.config import config, logger
+from common.exceptions import CCInvalidRequestException
 
-from handlers.utils import api_handler, authorize_compact
+from common.utils import api_handler, authorize_compact
 
 from . import get_provider_information
