+++ backend/compact-connect/lambdas/provider-data-v1/license_csv_reader.py
@@ -2,7 +2,7 @@ from collections.abc import Generator
 from csv import DictReader
 from io import TextIOBase
 
-from data_model.schema.license import LicensePostSchema
+from common.data_model.schema.license import LicensePostSchema
 
 
 class LicenseCSVReader:
