+++ backend/compact-connect/lambdas/provider-data-v1/tests/unit/test_data_model/test_schema/test_base_record.py
@@ -5,10 +5,10 @@ from tests import TstLambdas
 
 class TestRegistration(TstLambdas):
     def test_license_privilege_lookup(self):
-        from data_model.schema.base_record import BaseRecordSchema
-        from data_model.schema.license import LicenseRecordSchema
-        from data_model.schema.privilege import PrivilegeRecordSchema
-        from data_model.schema.provider import ProviderRecordSchema
+        from common.data_model.schema.base_record import BaseRecordSchema
+        from common.data_model.schema.license import LicenseRecordSchema
+        from common.data_model.schema.privilege import PrivilegeRecordSchema
+        from common.data_model.schema.provider import ProviderRecordSchema
 
         with open('tests/resources/dynamo/privilege.json') as f:
             privilege_data = json.load(f)
@@ -29,8 +29,8 @@ class TestRegistration(TstLambdas):
         self.assertIsInstance(provider_schema, ProviderRecordSchema)
 
     def test_invalid_type(self):
-        from data_model.schema.base_record import BaseRecordSchema
-        from exceptions import CCInternalException
+        from common.data_model.schema.base_record import BaseRecordSchema
+        from common.exceptions import CCInternalException
 
         with self.assertRaises(CCInternalException):
             BaseRecordSchema.get_schema_by_type('some-unsupported-type')
